Extended the input example description to include more discussion of how the input file assembles the equations that are solved. Added the neutron flux equations and indicated the kernel associated with each term.
TODO: I attempted to add the precursor concentration equations. However, there were a couple of terms that I had difficultly interpreting. Since I couldn't be sure of the form of the equation, I didn't include it. It would be vey helpful to document the Precursor action to clarify the form of the equations that are solved here.